fix(neis): NEIS 관련 데이터 동기화 시 데이터 유실 문제 수정

+++ src/main/kotlin/team/themoment/datagsm/domain/neis/meal/service/impl/SyncMealServiceImpl.kt
@@ -39,9 +39,10 @@ class SyncMealServiceImpl(
                 ?.map { convertToEntity(it) }
                 ?: emptyList()
 
-        val savedEntities = mealRedisRepository.saveAll(newMealEntities).toList()
-        mealRedisRepository.deleteAll()
-        mealRedisRepository.saveAll(savedEntities)
+        if (newMealEntities.isNotEmpty()) {
+            mealRedisRepository.deleteAll()
+            mealRedisRepository.saveAll(newMealEntities)
+        }
     }
 
     private fun convertToEntity(dto: MealServiceDietInfo): MealRedisEntity {
